Reorganized code and reused previously defined interfaces
