+++ packages/dumb/src/jest.config.js
@@ -158,7 +158,7 @@ module.exports = {
   // runner: "jest-runner",
 
   // The paths to modules that run some code to configure or set up the testing environment before each test
-  setupFiles: ['<rootDir>/beforeEach.js'],
+  setupFiles: ['<rootDir>/test/before-each.js'],
 
   // A list of paths to modules that run some code to configure or set up the testing framework before each test
   // setupFilesAfterEnv: [],
